+++ vrp-core/src/solver/evolution/run_simple.rs
@@ -1,4 +1,5 @@
 use crate::solver::evolution::*;
+use crate::solver::hyper::HyperHeuristic;
 use crate::solver::{RefinementContext, Telemetry};
 use crate::utils::Timer;
 
@@ -15,11 +16,12 @@ impl EvolutionStrategy for RunSimple {
     fn run(
         &self,
         refinement_ctx: RefinementContext,
-        mutation: &(dyn Mutation + Send + Sync),
+        hyper: Box<dyn HyperHeuristic + Send + Sync>,
         termination: &(dyn Termination + Send + Sync),
         telemetry: Telemetry,
     ) -> EvolutionResult {
         let mut refinement_ctx = refinement_ctx;
+        let mut hyper = hyper;
         let mut telemetry = telemetry;
 
         while !should_stop(&mut refinement_ctx, termination) {
@@ -27,7 +29,7 @@ impl EvolutionStrategy for RunSimple {
 
             let parents = refinement_ctx.population.select().collect();
 
-            let offspring = mutation.mutate_all(&refinement_ctx, parents);
+            let offspring = hyper.search(&refinement_ctx, parents);
 
             let is_improved =
                 if should_add_solution(&refinement_ctx) { refinement_ctx.population.add_all(offspring) } else { false };