+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- a new mutation operator: decompose search which is used for bigger problem instances
+- `breaking`: introduced hyper-heuristic model
+
 
 ## [v1.7.4] - 2021-01-23
 
@@ -11,9 +16,6 @@ All notable changes to this project will be documented in this file.
 
 - introduced parallelism control options (experimental).
 
-### Added
-
-- a new mutation operator: decompose search which is used for bigger problem instances
 
 ## [v1.7.3] - 2021-01-08